Time improvements
The main improvements are as follows:
1) Time.deltaTime() is divided by 1000 every frame instead of everytime it is gotten by the user; deltaTime is gotten a lot, especially in big games, and we don't want to repeat this operation when not needed.
2) Time.getTime() does not increment by 1000 every second, but instead by one. This allows the game to run much longer, and for timers to be much more accurate when needed.

+++ Time.h
@@ -19,16 +19,16 @@ public:
     //get game speed
 	static void setTimeScale(float newScale);
 	//get time elapsed since game start
-	static long getTime();
+	static double getTime();
 
 private:
 	static void calculateFPS();
 
 	static float _fps;
-	static float _maxFPS;
+	static int _maxFPS;
 	static float _deltaTime;
 	static float _timeScale;
-	static long _time;
+	static double _time;
 	static unsigned int _startTicks;
 };
 }
